media: add sipe_data_session_new()
Creating sipe_media_call_private using this function won't display
the usual Pidgin dialog for accepting/declining the call, and neither
the call-in-progress window with audio level meters and video will
show up. This will be used for Lync file transfer, where we want the
user to interact with our media session in a different way.